# Checksum WAL records