--- a/src/include/common/checksum.h
+++ b/src/include/common/checksum.h
@@ -1,0 +1,11 @@
+// From DuckDB
+#pragma once
+
+#include <cstddef>
+#include <cstdint>
+namespace kuzu::common {
+
+//! Compute a checksum over a buffer of size size
+uint64_t checksum(uint8_t* buffer, size_t size);
+
+} // namespace kuzu::common