--- a/src/include/common/checksum.h
+++ b/src/include/common/checksum.h
@@ -1,4 +1,3 @@
-// From DuckDB
 #pragma once
 
 #include <cstddef>