--- a/src/include/common/serializer/in_mem_file_writer.h
+++ b/src/include/common/serializer/in_mem_file_writer.h
@@ -34,7 +34,7 @@ public:
     page_idx_t getNumPagesToFlush() const { return pages.size(); }
 
     static uint64_t getPageSize();
-    void flush(BufferedFileWriter& writer) const;
+    void flush(Writer& writer) const;
 
     uint64_t getSize() const override {
         uint64_t size = pages.size() > 1 ? KUZU_PAGE_SIZE * (pages.size() - 1) : 0;