--- a/src/include/storage/wal/checksum_reader.h
+++ b/src/include/storage/wal/checksum_reader.h
@@ -3,6 +3,7 @@
 #include <memory>
 #include <optional>
 
+#include "common/serializer/deserializer.h"
 #include "common/serializer/reader.h"
 #include "storage/buffer_manager/memory_manager.h"
 
@@ -10,23 +11,22 @@ namespace kuzu {
 namespace storage {
 class ChecksumReader : public common::Reader {
 public:
-    explicit ChecksumReader(std::unique_ptr<Reader> reader, MemoryManager& memoryManager);
+    explicit ChecksumReader(common::FileInfo& fileInfo, MemoryManager& memoryManager);
 
     void read(uint8_t* data, uint64_t size) override;
     bool finished() override;
 
-    // Reads + stores the checksum for the next entry
-    void startEntry();
-    // Computes + verifies the checksum for the entry that has just been read
-    void finishEntry();
+    // Reads the stored checksum
+    // Also computes + verifies the checksum for the entry that has just been read against the
+    // stored value
+    void finishEntry(std::string_view checksumMismatchMessage);
 
-    Reader* getReader();
+    uint64_t getReadOffset() const;
 
 private:
-    std::unique_ptr<common::Deserializer> deserializer;
+    common::Deserializer deserializer;
 
     uint64_t currentEntrySize;
-    std::optional<uint64_t> currentChecksum;
     std::unique_ptr<MemoryBuffer> entryBuffer;
 };
 } // namespace storage