--- a/src/include/storage/wal/local_wal.h
+++ b/src/include/storage/wal/local_wal.h
@@ -1,5 +1,6 @@
 #pragma once
 
+#include "storage/wal/checksum_writer.h"
 #include "storage/wal/wal_record.h"
 
 namespace kuzu {
@@ -54,7 +55,8 @@ private:
 
 private:
     std::mutex mtx;
-    std::shared_ptr<common::InMemFileWriter> writer;
+    std::shared_ptr<ChecksumWriter> writer;
+    std::shared_ptr<common::InMemFileWriter> localWriter;
     std::unique_ptr<common::Serializer> serializer;
 };