--- a/src/include/storage/wal/wal.h
+++ b/src/include/storage/wal/wal.h
@@ -40,9 +40,12 @@ private:
     bool readOnly;
     common::VirtualFileSystem* vfs;
     std::unique_ptr<common::FileInfo> fileInfo;
+
+    // Since most writes to the shared WAL will be flushing local WAL (which has its own checksums),
+    // these writes can go through the normal writer We do still need a checksum writer though for
+    // writing COMMIT/CHECKPOINT records
     std::shared_ptr<common::BufferedFileWriter> writer;
-    std::shared_ptr<ChecksumWriter> checksumWriter;
-    std::unique_ptr<common::Serializer> serializer;
+    std::optional<ChecksumSerializer> checksumWriter;
 };
 
 } // namespace storage