--- a/src/include/storage/wal/wal_replayer.h
+++ b/src/include/storage/wal/wal_replayer.h
@@ -12,7 +12,7 @@ class WALReplayer {
 public:
     explicit WALReplayer(main::ClientContext& clientContext);
 
-    void replay() const;
+    void replay(bool throwOnWalReplayFailure) const;
 
 private:
     struct WALReplayInfo {
@@ -40,7 +40,7 @@ private:
 
     // This function is used to deserialize the WAL records without actually applying them to the
     // storage.
-    WALReplayInfo dryReplay(common::FileInfo& fileInfo) const;
+    WALReplayInfo dryReplay(common::FileInfo& fileInfo, bool throwOnWalReplayFailure) const;
 
     void removeWALAndShadowFiles() const;
     void removeFileIfExists(const std::string& path) const;