--- a/src/main/database.cpp
+++ b/src/main/database.cpp
@@ -34,7 +34,7 @@ namespace main {
 
 SystemConfig::SystemConfig(uint64_t bufferPoolSize_, uint64_t maxNumThreads, bool enableCompression,
     bool readOnly, uint64_t maxDBSize, bool autoCheckpoint, uint64_t checkpointThreshold,
-    bool forceCheckpointOnClose
+    bool forceCheckpointOnClose, bool throwOnWalReplayFailure
 #if defined(__APPLE__)
     ,
     uint32_t threadQos
@@ -42,7 +42,8 @@ SystemConfig::SystemConfig(uint64_t bufferPoolSize_, uint64_t maxNumThreads, boo
     )
     : maxNumThreads{maxNumThreads}, enableCompression{enableCompression}, readOnly{readOnly},
       autoCheckpoint{autoCheckpoint}, checkpointThreshold{checkpointThreshold},
-      forceCheckpointOnClose{forceCheckpointOnClose} {
+      forceCheckpointOnClose{forceCheckpointOnClose},
+      throwOnWalReplayFailure(throwOnWalReplayFailure) {
 #if defined(__APPLE__)
     this->threadQos = threadQos;
 #endif
@@ -132,7 +133,7 @@ void Database::initMembers(std::string_view dbPath, construct_bm_func_t initBmFu
         extensionManager->autoLoadLinkedExtensions(&clientContext);
         return;
     }
-    StorageManager::recover(clientContext);
+    StorageManager::recover(clientContext, dbConfig.throwOnWalReplayFailure);
 }
 
 Database::~Database() {