--- a/src/main/db_config.cpp
+++ b/src/main/db_config.cpp
@@ -10,7 +10,7 @@ namespace kuzu {
 namespace main {
 
 #define GET_CONFIGURATION(_PARAM)                                                                  \
-    {_PARAM::name, _PARAM::inputType, _PARAM::setContext, _PARAM::getSetting}
+    { _PARAM::name, _PARAM::inputType, _PARAM::setContext, _PARAM::getSetting }
 
 static ConfigurationOption options[] = { // NOLINT(cert-err58-cpp):
     GET_CONFIGURATION(ThreadsSetting), GET_CONFIGURATION(TimeoutSetting),