# Description

Add checksums to each individual record of the WAL, which we check against when replaying the WAL. This is done by making WAL reads/writes go through the new classes `ChecksumReader/Writer`. The functions `startEntry()`/`finishEntry()` are called before/after processing a WAL record, after which a buffer containing the accumulated serialized record in the reader/writer will be used to calculate the checksum.

## Performance

I also did a quick microbenchmark to get a feel for the performance impact of adding the checksum on replaying a transaction.

**Transaction:**

```
with kuzu.Database(tempdir + "/db.kz") as db, kuzu.Connection(db) as conn:
        conn.execute("call force_checkpoint_on_close=false")
        conn.execute("call auto_checkpoint=false")
        conn.execute("begin transaction")
        conn.execute("create node table nodes(id int64 primary key)")
        conn.execute("create rel table edges(from nodes to nodes)")

        NUM_NODES = 1000000
        conn.execute(f"unwind range(1, {NUM_NODES}) as i create (nodes{{id: i}})")
        conn.execute("commit")
```

|Branch|Average transaction execution time|Average replay time|
|--|--|--|
master|2.87s|2.40s
new|2.95s|2.54s

Average replay time on master: 1.23s
Average replay time on new branch: 1.34s

# Contributor agreement

- [ ] I have read and agree to the [Contributor Agreement](https://github.com/kuzudb/kuzu/blob/master/CLA.md).